typescript for the server
