+++ packages/server/src/comment/graphql/query.ts
@@ -1,5 +1,6 @@
-const { GraphQLNonNull, GraphQLInt, GraphQLList } = require('graphql');
-const { Comment } = require('./type');
+import { GraphQLNonNull, GraphQLInt, GraphQLList } from 'graphql';
+
+import { Comment } from './type';
 
 module.exports = {
   comment: {
@@ -9,7 +10,7 @@ module.exports = {
         type: new GraphQLNonNull(GraphQLInt),
       },
     },
-    resolve(parentValue, args) {
+    resolve(parentValue: any, args: any): any {
       // return axios
       //   .get(`${BASE_V1}/comments/${args.id}?client_id=${CLIENT_ID}`)
       //   .then(res => res.data);
@@ -32,7 +33,7 @@ module.exports = {
         defaultValue: 20,
       },
     },
-    resolve(parentValue, args) {
+    resolve(parentValue: any, args: any): any {
       // return axios
       //   .get(
       //     `${BASE_V1}/tracks/${args.trackId}/comments?limit=${
