+++ packages/server/src/comment/index.ts
@@ -1,10 +1,10 @@
-const queries = require('./graphql/query');
+import queries from './graphql/query';
 // import mutations from './graphql/mutation';
 
 /**
  * Setup GraphQL query fields for current module
  */
-function getGraphQLQuery() {
+export function getGraphQLQuery() {
   return {
     ...queries,
   };
@@ -18,7 +18,3 @@ function getGraphQLQuery() {
 //     ...mutations,
 //   };
 // }
-
-module.exports = {
-  getGraphQLQuery,
-};
