--- packages/server/src/common/env.js
@@ -1,5 +1,0 @@
-const CLIENT_ID = process.env.CLIENT_ID;
-
-module.exports = {
-  CLIENT_ID,
-};
