+++ packages/server/src/common/env.ts
@@ -1,0 +1,1 @@
+export const CLIENT_ID = process.env.CLIENT_ID;
