--- packages/server/src/common/soundcloud/consts.js
@@ -1,7 +1,0 @@
-const BASE_V1 = 'https://api.soundcloud.com';
-const BASE_V2 = 'https://api-v2.soundcloud.com';
-
-module.exports = {
-  BASE_V1,
-  BASE_V2,
-};
