+++ packages/server/src/common/soundcloud/consts.ts
@@ -1,0 +1,2 @@
+export const BASE_V1 = 'https://api.soundcloud.com';
+export const BASE_V2 = 'https://api-v2.soundcloud.com';
