--- packages/server/src/config/logger.js
@@ -1,8 +1,0 @@
-function configureLogger(app) {
-  app.use((req, res, next) => {
-    console.log(new Date(), req.method, req.url);
-    next();
-  });
-}
-
-module.exports = configureLogger;
