--- packages/server/src/track/service.js
@@ -1,9 +1,0 @@
-const trackRepo = require('./repository');
-
-function getTrackById(trackId) {
-  return trackRepo.getTrackById(trackId);
-}
-
-module.exports = {
-  getTrackById,
-};
