+++ packages/server/src/track/service.ts
@@ -1,0 +1,5 @@
+import * as trackRepo from './repository';
+
+export function getTrackById(trackId: number) {
+  return trackRepo.getTrackById(trackId);
+}
