+++ packages/server/src/user/graphql/query.ts
@@ -1,5 +1,5 @@
-const { GraphQLNonNull, GraphQLInt } = require('graphql');
-const { User } = require('./type');
+import { GraphQLNonNull, GraphQLInt } from 'graphql';
+import { User } from './type';
 
 module.exports = {
   user: {
@@ -9,7 +9,7 @@ module.exports = {
         type: new GraphQLNonNull(GraphQLInt),
       },
     },
-    resolve(parentValue, args) {
+    resolve(parentValue: any, args: any): any {
       // return axios.get(`${BASE_V1}/users/${args.id}?client_id=${CLIENT_ID}`).then(res => res.data);
       return null;
     },
